+++ vscp/test/vscpUser/vscp_portable.h
@@ -59,7 +59,7 @@ This module contains VSCP support functionality, which is used by the core.
 /*******************************************************************************
     INCLUDES
 *******************************************************************************/
-#include <inttypes.h>
+#include <stdint.h>
 #include "vscp_types.h"
 #include "vscp_config.h"